+++ docs/docs/en/api/fastagency/app/validate_model.md
@@ -8,4 +8,4 @@ search:
   boost: 0.5
 ---
 
-::: fastagency.models.agents.agents.AssistantAgent
+::: fastagency.app.validate_model
