+++ docs/docs/en/api/fastagency/models/ModelSchemas.md
@@ -8,4 +8,4 @@ search:
   boost: 0.5
 ---
 
-::: fastagency.app.validate_llm_model
+::: fastagency.models.ModelSchemas
