+++ docs/docs/en/api/fastagency/models/agents/WebSurferAgent.md
@@ -1,0 +1,11 @@
+---
+# 0.5 - API
+# 2 - Release
+# 3 - Contributing
+# 5 - Template Page
+# 10 - Default
+search:
+  boost: 0.5
+---
+
+::: fastagency.models.agents.WebSurferAgent
