+++ fastagency/models/agents/__init__.py
@@ -1,0 +1,4 @@
+from ._assistant import AssistantAgent
+from ._web_surfer import WebSurferAgent
+
+__all__ = ["AssistantAgent", "WebSurferAgent"]
