+++ tests/app/test_openai_extensively.py
@@ -1,9 +1,7 @@
-import pytest
 from fastapi.testclient import TestClient
 
 from fastagency.app import app
 from fastagency.models import ModelSchemas, Schemas
-from fastagency.models.agents.agents import AssistantAgent, WebSurferAgent
 from fastagency.models.llms import OpenAI
 
 client = TestClient(app)
@@ -180,166 +178,3 @@ def test_get_schemas() -> None:
 
     schemas = Schemas(**response.json())
     assert len(schemas.schemas) >= 2
-
-
-@pytest.mark.skip(reason="Not implemented yet")
-class TestAgents:
-    def test_list_agents(self) -> None:
-        response = client.get("/models/agents")
-        assert response.status_code == 200
-        expected = ["AssistantAgent", "WebSurferAgent"]
-        assert response.json() == expected
-
-
-@pytest.mark.skip(reason="Not implemented yet")
-class TestAssistantAgents:
-    def test_get_assistant_agent_schema(self) -> None:
-        response = client.get(f"/models/agents/{AssistantAgent.__name__}")
-        assert response.status_code == 200
-        expected = {
-            "properties": {
-                "uuid": {
-                    "description": "The unique identifier",
-                    "format": "uuid",
-                    "title": "UUID",
-                    "type": "string",
-                },
-                "name": {
-                    "description": "The name of the agent",
-                    "title": "Name",
-                    "type": "string",
-                },
-                "llm_uuid": {
-                    "description": "The unique identifier for the LLM instance",
-                    "format": "uuid",
-                    "title": "LLM UUID",
-                    "type": "string",
-                },
-                "system_message": {
-                    "description": "The system message of the agent. This message is used to inform the agent about his role in the conversation",
-                    "title": "System Message",
-                    "type": "string",
-                },
-            },
-            "required": ["uuid", "name", "llm_uuid", "system_message"],
-            "title": "AssistantAgent",
-            "type": "object",
-        }
-        assert response.json() == expected
-
-    def test_validate_assistant_agent_success(self) -> None:
-        response = client.post(
-            f"/models/agents/{AssistantAgent.__name__}/validate",
-            json={
-                "uuid": "12345678-1234-5678-1234-567812345678",
-                "llm_uuid": "87654321-1234-5678-1234-567812345678",
-                "name": "test agent",
-                "system_message": "test system message",
-            },
-        )
-        assert response.status_code == 200
-
-    def test_validate_assistant_agent_missing_llm_uuid(self) -> None:
-        response = client.post(
-            f"/models/agents/{AssistantAgent.__name__}/validate",
-            json={
-                "uuid": "12345678-1234-5678-1234-567812345678",
-                "name": "test agent",
-                "system_message": "test system message",
-            },
-        )
-        assert response.status_code == 422
-        msg_dict = response.json()["detail"][0]
-        msg_dict.pop("input")
-        # msg_dict.pop("url")
-        excepted = {
-            "type": "missing",
-            "loc": ["body", "llm_uuid"],
-            "msg": "Field required",
-        }
-        assert msg_dict == excepted
-
-    def test_validate_assistant_agent_invalid_llm_uuid(self) -> None:
-        response = client.post(
-            f"/models/agents/{AssistantAgent.__name__}/validate",
-            json={
-                "llm_uuid": "87654321-1234-5678",
-                "uuid": "12345678-1234-5678-1234-567812345678",
-                "name": "test agent",
-                "system_message": "test system message",
-            },
-        )
-        assert response.status_code == 422
-        msg_dict = response.json()["detail"][0]
-        msg_dict.pop("input")
-        # msg_dict.pop("url")
-        excepted = {
-            "type": "uuid_parsing",
-            "loc": ["body", "llm_uuid"],
-            "msg": "Input should be a valid UUID, invalid group count: expected 5, found 3",
-            "ctx": {"error": "invalid group count: expected 5, found 3"},
-        }
-        assert msg_dict == excepted
-
-
-@pytest.mark.skip(reason="Not implemented yet")
-class TestWebSurferAgent:
-    def test_get_web_surfer_agent_schema(self) -> None:
-        response = client.get(f"/models/agents/{WebSurferAgent.__name__}")
-        assert response.status_code == 200
-        expected = {
-            "properties": {
-                "uuid": {
-                    "description": "The unique identifier",
-                    "format": "uuid",
-                    "title": "UUID",
-                    "type": "string",
-                },
-                "name": {
-                    "description": "The name of the agent",
-                    "title": "Name",
-                    "type": "string",
-                },
-                "llm_uuid": {
-                    "description": "The unique identifier for the LLM instance",
-                    "format": "uuid",
-                    "title": "LLM UUID",
-                    "type": "string",
-                },
-                "summarizer_llm_uuid": {
-                    "description": "The unique identifier for the summarizer model instance",
-                    "format": "uuid",
-                    "title": "Summarizer LLM UUID",
-                    "type": "string",
-                },
-                "viewport_size": {
-                    "default": 1080,
-                    "description": "The viewport size of the browser",
-                    "title": "Viewport Size",
-                    "type": "integer",
-                },
-                "bing_api_key": {
-                    "anyOf": [{"type": "string"}, {"type": "null"}],
-                    "default": None,
-                    "description": "The Bing API key for the browser",
-                    "title": "Bing Api Key",
-                },
-            },
-            "required": ["uuid", "name", "llm_uuid", "summarizer_llm_uuid"],
-            "title": "WebSurferAgent",
-            "type": "object",
-        }
-        assert response.json() == expected
-
-    def test_validate_web_surfer_agent_success(self) -> None:
-        response = client.post(
-            f"/models/agents/{WebSurferAgent.__name__}/validate",
-            json={
-                "uuid": "12345678-1234-5678-1234-567812345678",
-                "llm_uuid": "87654321-1234-5678-1234-567812345678",
-                "summarizer_llm_uuid": "87654321-1234-5678-1234-567812345678",
-                "name": "test agent",
-                "system_message": "test system message",
-            },
-        )
-        assert response.status_code == 200
